requirements.txt
# # Core AI and ML libraries
# openai>=1.0.0
# anthropic>=0.8.0
# cohere>=4.0.0
# ollama>=0.4.0

# # Document processing
# python-docx>=0.8.11
# markdown>=3.4.0
# beautifulsoup4>=4.11.0
# pypandoc>=1.11

# # Configuration and data handling
# pyyaml>=6.0
# requests>=2.28.0
# pandas>=1.5.0
# numpy>=1.24.0

# # Web scraping and data extraction
# scrapy>=2.8.0
# selenium>=4.8.0
# requests-html>=0.10.0

# # Bitcoin SV specific libraries
# bitcoinx>=0.9
# python-bitcointx>=1.1.5

# # MCP (Model Context Protocol) integration
# mcp-client>=0.1.0
# websockets>=11.0

# # Database and caching
# redis>=4.5.0

# # Utility libraries
# click>=8.1.0
# tqdm>=4.64.0
# python-dotenv>=1.0.0
# loguru>=0.7.0

# # Settings and validation
# pydantic>=2.0.0

# Core AI libraries
openai>=1.0.0
anthropic>=0.8.0
cohere>=4.0.0
ollama>=0.4.0

# Document processing
python-docx>=0.8.11
markdown>=3.4.0
beautifulsoup4>=4.11.0

# Configuration and data handling
pyyaml>=6.0
requests>=2.28.0

# Utility libraries
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0

